requirements.txt
# AI Financial Advisor V2 - Requirements
# Core Framework
flask==3.0.3
flask-cors==4.0.1
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0

# Frontend
streamlit==1.38.0

# AI/ML
google-genai==1.0.0

# Data Processing
pandas>=2.2.3
numpy>=2.1.1

# Visualization
plotly==5.24.1
matplotlib>=3.9.2

# Configuration
python-dotenv==1.0.1

# Validation
pydantic==2.9.2
marshmallow==3.22.0

# HTTP
requests==2.32.3

# Database
sqlalchemy==2.0.35
psycopg2-binary>=2.9.9
alembic==1.13.3

# Security
bcrypt==4.2.0

# Testing
pytest==8.3.3
pytest-flask==1.3.0

# Documentation
flasgger>=0.9.5
